feat: 異変の抽選処理を追加した
