+++ app/routes/problems.tsx
@@ -75,24 +75,13 @@ const Problems: React.FC = () => {
       }
     })();
 
-    // DOM 更新処理
-    const elm = document.querySelector("header");
-    if (!elm) {
-      console.error("Element not found:");
-      return;
+    // 異変の抽選処理
+    if (Math.random() < 0.6) {
+      (async () => {
+        const anomaly = await import("../resources/anomalies/rainbow-header");
+        anomaly.default.execute();
+      })();
     }
-    const styleElement = document.createElement("style");
-    styleElement.textContent = `
-      .rainbow-gradient {
-        background: linear-gradient(to right,#e60000,#f39800,#fff100,#009944,#0068b7,#1d2088,#920783,#e60000) 0 / 200%;
-        animation: 5s example4 linear infinite;
-      }
-      @keyframes example4 {
-        100% { background-position: 200%; }
-      }
-    `;
-    document.head.appendChild(styleElement);
-    elm.classList.add("rainbow-gradient");
   }, []);
 
   if (error) {
